- heavy renaming / refactoring - adding an endpoint to get recordings for a user. this request works with paging and is prepared for a search term, but this is not implemented
