--- backend/src/main/java/com/github/mysterix5/vover/model/Accessibility.java
@@ -1,5 +1,0 @@
-package com.github.mysterix5.vover.model;
-
-public enum Accessibility {
-    ONLY_FOR_YOURSELF, FRIENDS, PUBLIC
-}
