--- backend/src/main/java/com/github/mysterix5/vover/model/Availability.java
@@ -1,5 +1,0 @@
-package com.github.mysterix5.vover.model;
-
-public enum Availability {
-    INVALID, ABSENT, SCOPE, PUBLIC
-}
