--- backend/src/main/java/com/github/mysterix5/vover/model/TextSubmitDTO.java
@@ -1,8 +1,0 @@
-package com.github.mysterix5.vover.model;
-
-import lombok.Data;
-
-@Data
-public class TextSubmitDTO {
-    private String text;
-}
