--- backend/src/main/java/com/github/mysterix5/vover/model/UserRegisterDTO.java
@@ -1,14 +1,0 @@
-package com.github.mysterix5.vover.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserRegisterDTO {
-    private String username;
-    private String password;
-    private String passwordRepeat;
-}
