--- backend/src/main/java/com/github/mysterix5/vover/model/VoverUser.java
@@ -1,20 +1,0 @@
-package com.github.mysterix5.vover.model;
-
-import lombok.Data;
-import org.springframework.data.annotation.Id;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Data
-public class VoverUser {
-    @Id
-    private String id;
-    private String username;
-    private String password;
-    private List<String> roles = new ArrayList<>();
-
-    public void addRole(String role){
-        roles.add(role);
-    }
-}
