--- backend/src/main/java/com/github/mysterix5/vover/model/WordResponseDTO.java
@@ -1,19 +1,0 @@
-package com.github.mysterix5.vover.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class WordResponseDTO {
-    private String word;
-    private Availability availability;
-
-    public WordResponseDTO(String word){
-        this.word = word;
-    }
-}
