+++ backend/src/main/java/com/github/mysterix5/vover/model/other/Accessibility.java
@@ -1,0 +1,5 @@
+package com.github.mysterix5.vover.model.other;
+
+public enum Accessibility {
+    ONLY_FOR_YOURSELF, FRIENDS, PUBLIC
+}
