+++ backend/src/main/java/com/github/mysterix5/vover/model/other/Availability.java
@@ -1,0 +1,5 @@
+package com.github.mysterix5.vover.model.other;
+
+public enum Availability {
+    INVALID, ABSENT, SCOPE, PUBLIC
+}
