+++ backend/src/main/java/com/github/mysterix5/vover/model/security/UserAuthenticationDTO.java
@@ -1,0 +1,13 @@
+package com.github.mysterix5.vover.model.security;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserAuthenticationDTO {
+    private String username;
+    private String password;
+}
