+++ backend/src/main/java/com/github/mysterix5/vover/model/security/UserRegisterDTO.java
@@ -1,0 +1,14 @@
+package com.github.mysterix5.vover.model.security;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserRegisterDTO {
+    private String username;
+    private String password;
+    private String passwordRepeat;
+}
