+++ backend/src/main/java/com/github/mysterix5/vover/model/security/VoverUserEntity.java
@@ -1,0 +1,22 @@
+package com.github.mysterix5.vover.model.security;
+
+import lombok.Data;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Data
+@Document(collection = "voverUser")
+public class VoverUserEntity {
+    @Id
+    private String id;
+    private String username;
+    private String password;
+    private List<String> roles = new ArrayList<>();
+
+    public void addRole(String role){
+        roles.add(role);
+    }
+}
