+++ backend/src/main/java/com/github/mysterix5/vover/model/text/TextSubmitDTO.java
@@ -1,0 +1,8 @@
+package com.github.mysterix5.vover.model.text;
+
+import lombok.Data;
+
+@Data
+public class TextSubmitDTO {
+    private String text;
+}
