+++ backend/src/main/java/com/github/mysterix5/vover/model/word/RecordPage.java
@@ -1,0 +1,16 @@
+package com.github.mysterix5.vover.model.word;
+
+import lombok.Builder;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+@Builder
+public class RecordPage {
+    private int page;
+    private int noPages;
+    private int size;
+    private String searchTerm;
+    private List<RecordManagementDTO> records;
+}
