+++ backend/src/main/java/com/github/mysterix5/vover/model/word/RecordPageSubmitDTO.java
@@ -1,0 +1,10 @@
+package com.github.mysterix5.vover.model.word;
+
+import lombok.Data;
+
+@Data
+public class RecordPageSubmitDTO {
+    private int page;
+    private int size;
+    private String searchTerm;
+}
