--- backend/src/main/java/com/github/mysterix5/vover/textHandling/WordsMongoRepository.java
@@ -1,11 +1,0 @@
-package com.github.mysterix5.vover.textHandling;
-
-import com.github.mysterix5.vover.model.WordDbEntity;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import java.util.List;
-import java.util.Set;
-
-public interface WordsMongoRepository extends MongoRepository<WordDbEntity, String> {
-    List<WordDbEntity> findByWordIn(Set<String> words);
-}
