+++ backend/src/main/java/com/github/mysterix5/vover/usermanagement/UserMongoRepository.java
@@ -1,13 +1,13 @@
 package com.github.mysterix5.vover.usermanagement;
 
-import com.github.mysterix5.vover.model.VoverUser;
+import com.github.mysterix5.vover.model.security.VoverUserEntity;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserMongoRepository extends MongoRepository<VoverUser, String> {
-    Optional<VoverUser> findByUsername(String username);
+public interface UserMongoRepository extends MongoRepository<VoverUserEntity, String> {
+    Optional<VoverUserEntity> findByUsername(String username);
     boolean existsByUsername(String username);
 }
