+++ backend/src/main/java/com/github/mysterix5/vover/words/WordController.java
@@ -1,0 +1,49 @@
+package com.github.mysterix5.vover.words;
+
+import com.github.mysterix5.vover.model.other.VoverErrorDTO;
+import com.github.mysterix5.vover.model.word.RecordPage;
+import com.github.mysterix5.vover.model.word.RecordPageSubmitDTO;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.security.Principal;
+
+@Slf4j
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/api/word")
+public class WordController {
+    private final WordService wordService;
+
+    @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+    public ResponseEntity<Object> addWord(@RequestParam("word") String word,
+                                        @RequestParam("tag") String tag,
+                                        @RequestParam("accessibility") String accessibility,
+                                        @RequestParam("audio") MultipartFile audio,
+                                        Principal principal
+    ) {
+        try {
+            var audioBytes = audio.getBytes();
+            wordService.addWordToDb(word.toLowerCase(), principal.getName(), tag.toLowerCase(), accessibility, audioBytes);
+        }catch(Exception e){
+            return ResponseEntity.internalServerError().body(new VoverErrorDTO("Something went wrong while saving your recording"));
+        }
+
+        return ResponseEntity.ok().build();
+    }
+
+    @GetMapping
+    public ResponseEntity<Object> getRecordPage(@RequestBody RecordPageSubmitDTO recordPageSubmitDTO, Principal principal){
+        try{
+            RecordPage RecordPage = wordService.getRecordPage(principal.getName(), recordPageSubmitDTO.getPage(), recordPageSubmitDTO.getSize(), recordPageSubmitDTO.getSearchTerm());
+            return ResponseEntity.ok().body(RecordPage);
+        }catch (Exception e){
+            return ResponseEntity.internalServerError().body(new VoverErrorDTO("Something went wrong fetching your records"));
+        }
+    }
+
+}
